Move wrappers into its own module
This avoids polluting the module with system includes that can affect
how the wrapping function declarations are done, via remapping or
aliases that are outside our control.

This fixes building the project with LFS and time64 support enable.

Fixes: #54
